+++ frontend/src/lib/components/FlowPreviewContent.svelte
@@ -184,7 +184,9 @@
 			const previousJobId = await JobService.listJobs({
 				workspace: $workspaceStore!,
 				scriptPathExact:
-					($initialPathStore == '' ? $pathStore : $initialPathStore) + '/' + module.id,
+					`path` in module.value
+						? module.value.path
+						: ($initialPathStore == '' ? $pathStore : $initialPathStore) + '/' + module.id,
 				jobKinds: ['preview', 'script', 'flowpreview', 'flow'].join(','),
 				page: 1,
 				perPage: 1
@@ -196,10 +198,13 @@
 					workspace: $workspaceStore!,
 					id: previousJobId[0].id
 				})
-				if (getJobResult.result) {
+				if ('result' in getJobResult) {
 					$flowStateStore[module.id] = {
 						...($flowStateStore[module.id] ?? {}),
-						previewResult: getJobResult.result
+						previewResult: getJobResult.result,
+						previewJobId: previousJobId[0].id,
+						previewWorkspaceId: previousJobId[0].workspace_id,
+						previewSuccess: getJobResult.success
 					}
 				}
 			}
@@ -363,7 +368,7 @@
 					class="bg-orange-200 text-orange-600 border border-orange-600 p-2 flex items-center gap-2 rounded"
 				>
 					<AlertTriangle size={14} /> Flow changed since last preview
-					<div class="flex" />
+					<div class="flex"></div>
 				</div>
 			</div>
 		{/if}